# Use concrete array to enumerate job listeners